+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Changed
 
+- All the modules in [`pkg/sdk`](https://github.com/operator-framework/operator-sdk/tree/master/pkg/sdk) have been combined into a single package. `action`, `handler`, `informer` `types` and `query` pkgs have been consolidated into `pkg/sdk`. [#242](https://github.com/operator-framework/operator-sdk/pull/242)
+
 ### Removed
 
 ### Fixed